[Refactor] Move chunk merge utilities to ComputeEnv
