+++ be/module_boundary_manifest.json
@@ -342,7 +342,6 @@
         "runtime/schema_table_sink.h",
         "runtime/small_file_mgr.h",
         "runtime/snapshot_loader.h",
-        "runtime/sorted_chunks_merger.h",
         "runtime/table_function_table_sink.h",
         "runtime/tablets_channel.h",
         "runtime/variable_result_writer.h"
