+++ be/src/bench/chunks_sorter_bench.cpp
@@ -26,15 +26,14 @@
 #include "common/runtime_profile.h"
 #include "compute_env/sorting/merge.h"
 #include "compute_env/sorting/sort_permute.h"
+#include "compute_env/sorting/sorted_chunks_merger.h"
 #include "compute_env/sorting/sorting.h"
 #include "exec/chunks_sorter.h"
 #include "exec/chunks_sorter_full_sort.h"
 #include "exec/chunks_sorter_heap_sort.h"
 #include "exec/chunks_sorter_topn.h"
 #include "exprs/column_ref.h"
-#include "runtime/chunk_cursor.h"
 #include "runtime/runtime_state.h"
-#include "runtime/sorted_chunks_merger.h"
 #include "types/logical_type.h"
 #include "types/type_descriptor.h"
