+++ be/src/exec/pipeline/exchange/exchange_merge_sort_source_operator.cpp
@@ -15,13 +15,13 @@
 #include "exec/pipeline/exchange/exchange_merge_sort_source_operator.h"
 
 #include "common/config_exec_flow_fwd.h"
+#include "compute_env/sorting/sorted_chunks_merger.h"
 #include "exec/sort_exec_exprs.h"
 #include "runtime/data_stream_mgr.h"
 #include "runtime/data_stream_recvr.h"
 #include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
 #include "runtime/runtime_state_helper.h"
-#include "runtime/sorted_chunks_merger.h"
 
 namespace starrocks::pipeline {
 Status ExchangeMergeSortSourceOperator::prepare(RuntimeState* state) {
