+++ be/src/runtime/data_stream_recvr.cpp
@@ -47,17 +47,16 @@
 #include "common/runtime_profile.h"
 #include "common/util/debug_util.h"
 #include "compute_env/sorting/sort_cursor.h"
+#include "compute_env/sorting/sorted_chunks_merger.h"
 #include "exec/pipeline/query_context.h"
 #include "exec/sort_exec_exprs.h"
 #include "gen_cpp/data.pb.h"
 #include "gen_cpp/internal_service.pb.h"
-#include "runtime/chunk_cursor.h"
 #include "runtime/current_thread.h"
 #include "runtime/data_stream_mgr.h"
 #include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
 #include "runtime/sender_queue.h"
-#include "runtime/sorted_chunks_merger.h"
 #include "util/logging.h"
 
 namespace starrocks {
